[bulkidx] R3-04: fix release -Werror sign-compare/parentheses in bulk recovery
Release preset enables -Werror; two pre-existing R3-04 sites broke the
release build (debug did not flag them):
- btree.c:32317 btree_bulk_unique_absent_repair: int rv_redo_data_length vs
  size_t BTREE_RV_BUFFER_SIZE (cast to size_t, matching the file's idiom).
- file_manager.c file_recovery_check_vpid: INT16 offset/max_size vs size_t
  (DB_PAGESIZE - sizeof-based ALIGNED_SIZE) comparisons (cast RHS to int).
No behavior change; validation semantics identical.

+++ src/storage/btree.c
@@ -32314,7 +32314,7 @@ btree_bulk_unique_absent_repair (THREAD_ENTRY * thread_p, BTID_INT * btid_int, D
 
   /* Preflight before the first page mutation. */
   BTREE_RV_GET_DATA_LENGTH (rv_redo_data_ptr, rv_redo_data, rv_redo_data_length);
-  if (rv_redo_data_length > BTREE_RV_BUFFER_SIZE
+  if ((size_t) rv_redo_data_length > BTREE_RV_BUFFER_SIZE
       || (leaf_record->length > old_length
 	  && spage_get_free_space_without_saving (thread_p, leaf_page, NULL)
 	  < leaf_record->length - old_length))
